update(control-panels): optimized structure
